Sufficiently updated docs, build system and a few of the examples for an initial release.
Primary remaining problem: udev-style kernels don't allow registration of a
preexisting class right now.  This will need to be fixed semi-immediately.



git-svn-id: http://svn.xiph.org/trunk/fusd@12316 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ examples/drums3.c
@@ -43,7 +43,7 @@
  * However, it also prints a prompt to the console, asking the user if
  * how loud the drums should be.
  *
- * $Id: drums3.c,v 1.3 2003/07/11 22:29:38 cerpa Exp $
+ * $Id$
  */
 
 #include <stdio.h>
@@ -63,7 +63,7 @@ static char *drums_strings[] = {"bam", "bum", "beat", "boom",
 
 int volume = 2; /* default volume is 2 */
 
-int drums_read(struct fusd_file_info *file, char *user_buffer,
+ssize_t drums_read(struct fusd_file_info *file, char *user_buffer,
 	       size_t user_length, loff_t *offset)
 {
   int len;
